Voice Chat Info
Added the type VoiceChatStarted and the field voice_chat_started to the class Message.
Added the type VoiceChatEnded and the field voice_chat_ended to the class Message.
Added the type VoiceChatParticipantsInvited and the field voice_chat_participants_invited to the class Message.
Added the new administrator privilege can_manage_voice_chats to the class ChatMember and parameter can_manage_voice_chats to the method promoteChatMember. For now, bots can use this privilege only for passing to other administrators.