Improved Invite Links
Added the class ChatInviteLink, representing an invite link to a chat.
Added the method createChatInviteLink, which can be used to create new invite links in addition to the primary invite link.
Added the method editChatInviteLink, which can be used to edit non-primary invite links created by the bot.
Added the method revokeChatInviteLink, which can be used to revoke invite links created by the bot.